Better escaping for both
