+++ src/palace/manager/service/redis/models/marc.py
@@ -9,9 +9,9 @@ from functools import cached_property
 from typing import Any
 
 from pydantic import BaseModel
-from redis import ResponseError, WatchError
+from redis import WatchError
 
-from palace.manager.core.exceptions import PalaceValueError
+from palace.manager.service.redis.escape import JsonPathEscapeMixin
 from palace.manager.service.redis.models.lock import LockError, RedisJsonLock
 from palace.manager.service.redis.redis import Pipeline, Redis
 from palace.manager.service.storage.s3 import MultipartS3UploadPart
@@ -41,85 +41,7 @@ class MarcFileUploadState(StrEnum):
     UPLOADING = auto()
 
 
-class PathEscapeMixin:
-    """
-    Mixin to provide methods for escaping and unescaping paths for use in redis.
-
-    This is necessary because it seems like there is a bug in the AWS elasticache implementation
-    of JSONPATH where slashes or tilde character within a string literal used as a key cause issues.
-    This bug is not present in the open source redis implementation, which does the sane thing, not
-    requiring any special escaping.
-
-    Hopefully at some point AWS will fix these issues, and we can drop this mixin, so I tried to
-    encapsulate the logic for this here.
-
-    In AWS when a tilde is used in a key, the key is never updated, despite returning a success. And
-    when a slash is used in a key, the key is interpreted as a nested path, nesting a new key for every
-    slash in the path. This is not the behavior we want, so we need to escape these characters.
-
-    We can test if this is fixed in the future by running the test suite against AWS elasticache with
-    this mixin removed. If the tests pass, then it can be removed.
-
-    Characters are escaped by prefixing them with a backtick character, followed by a single character
-    from _MAPPING that represents the escaped character. The backtick character itself is escaped by
-    prefixing it with another backtick character.
-    """
-
-    _ESCAPE_CHAR = "`"
-
-    _MAPPING = {
-        "/": "s",
-        "~": "t",
-    }
-
-    @cached_property
-    def _FORWARD_MAPPING(self) -> dict[str, str]:
-        mapping = {k: "".join((self._ESCAPE_CHAR, v)) for k, v in self._MAPPING.items()}
-        mapping[self._ESCAPE_CHAR] = "".join((self._ESCAPE_CHAR, self._ESCAPE_CHAR))
-        return mapping
-
-    @cached_property
-    def _REVERSE_MAPPING(self) -> dict[str, str]:
-        mapping = {v: k for k, v in self._MAPPING.items()}
-        mapping[self._ESCAPE_CHAR] = self._ESCAPE_CHAR
-        return mapping
-
-    def _escape_path(self, path: str) -> str:
-        escaped = json.dumps("".join([self._FORWARD_MAPPING.get(c, c) for c in path]))
-        return escaped[1:-1]
-
-    def _unescape_path(self, path: str) -> str:
-        # Normal redis paths are always double-quoted, so we can use json.loads to unescape them.
-        # This does not happen in the AWS elasticache implementation, so we need to handle it manually,
-        # so that we can support both implementations.
-        try:
-            path = json.loads(f'"{path}"')
-        except json.JSONDecodeError:
-            pass
-
-        in_escape = False
-        unescaped = []
-        for char in path:
-            if in_escape:
-                if char not in self._REVERSE_MAPPING:
-                    raise PalaceValueError(
-                        f"Invalid escape sequence '{self._ESCAPE_CHAR}{char}'"
-                    )
-                unescaped.append(self._REVERSE_MAPPING[char])
-                in_escape = False
-            else:
-                if char == self._ESCAPE_CHAR:
-                    in_escape = True
-                else:
-                    unescaped.append(char)
-
-        if in_escape:
-            raise PalaceValueError("Unterminated escape sequence.")
-
-        return "".join(unescaped)
-
-
-class MarcFileUploadSession(RedisJsonLock, PathEscapeMixin, LoggerMixin):
+class MarcFileUploadSession(RedisJsonLock, JsonPathEscapeMixin, LoggerMixin):
     """
     This class is used as a lock for the Celery MARC export task, to ensure that only one
     task can upload MARC files for a given collection at a time. It increments an update
@@ -185,7 +107,7 @@ class MarcFileUploadSession(RedisJsonLock, PathEscapeMixin, LoggerMixin):
         return MarcFileUpload(buffer=buffer_data).dict(exclude_none=True)
 
     def _upload_path(self, upload_key: str) -> str:
-        upload_key = self._escape_path(upload_key)
+        upload_key = self._escape_path(upload_key, self._redis_client.elasticache)
         return f'{self._uploads_json_key}["{upload_key}"]'
 
     def _buffer_path(self, upload_key: str) -> str:
@@ -255,18 +177,6 @@ class MarcFileUploadSession(RedisJsonLock, PathEscapeMixin, LoggerMixin):
 
         return pipe_results[:-3]
 
-    @staticmethod
-    def _validate_results(results: list[Any]) -> bool:
-        """
-        This function validates that all the results of the pipeline are successful,
-        and not a ResponseError.
-
-        NOTE: The AWS elasticache implementation returns slightly different results then redis.
-        In redis, unsuccessful results when a key is not found are `None`, but in AWS they are
-        returned as a `ResponseError`, which is why we are checking for both in this function.
-        """
-        return all(r and not isinstance(r, ResponseError) for r in results)
-
     def append_buffers(self, data: Mapping[str, str]) -> dict[str, int]:
         if not data:
             return {}
@@ -295,7 +205,7 @@ class MarcFileUploadSession(RedisJsonLock, PathEscapeMixin, LoggerMixin):
 
             pipe_results = self._execute_pipeline(pipe, len(data))
 
-        if not self._validate_results(pipe_results):
+        if not self._validate_pipeline_results(pipe_results):
             raise MarcFileUploadSessionError("Failed to append buffers.")
 
         return {
@@ -317,7 +227,7 @@ class MarcFileUploadSession(RedisJsonLock, PathEscapeMixin, LoggerMixin):
             )
             pipe_results = self._execute_pipeline(pipe, 1)
 
-        if not self._validate_results(pipe_results):
+        if not self._validate_pipeline_results(pipe_results):
             raise MarcFileUploadSessionError("Failed to add part and clear buffer.")
 
     def set_upload_id(self, key: str, upload_id: str) -> None:
@@ -330,7 +240,7 @@ class MarcFileUploadSession(RedisJsonLock, PathEscapeMixin, LoggerMixin):
             )
             pipe_results = self._execute_pipeline(pipe, 1)
 
-        if not self._validate_results(pipe_results):
+        if not self._validate_pipeline_results(pipe_results):
             raise MarcFileUploadSessionError("Failed to set upload ID.")
 
     def clear_uploads(self) -> None:
@@ -338,7 +248,7 @@ class MarcFileUploadSession(RedisJsonLock, PathEscapeMixin, LoggerMixin):
             pipe.json().clear(self.key, self._uploads_json_key)
             pipe_results = self._execute_pipeline(pipe, 1)
 
-        if not self._validate_results(pipe_results):
+        if not self._validate_pipeline_results(pipe_results):
             raise MarcFileUploadSessionError("Failed to clear uploads.")
 
     def _get_specific(
@@ -382,7 +292,7 @@ class MarcFileUploadSession(RedisJsonLock, PathEscapeMixin, LoggerMixin):
             pipe.json().get(self.key, self._buffer_path(key))
             pipe.json().arrlen(self.key, self._parts_path(key))
             results = pipe.execute(raise_on_error=False)
-        if not self._validate_results(results):
+        if not self._validate_pipeline_results(results):
             raise MarcFileUploadSessionError(
                 "Failed to get part number and buffer data."
             )
