+++ src/palace/manager/service/redis/redis.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
+from functools import cached_property
 from typing import TYPE_CHECKING, Any
 
 import redis
@@ -76,6 +77,7 @@ class RedisPrefixCheckMixin(ABC):
         cmd.name: cmd
         for cmd in [
             RedisCommandNoArgs("SCRIPT LOAD"),
+            RedisCommandNoArgs("INFO"),
             RedisCommandArgs("KEYS"),
             RedisCommandArgs("GET"),
             RedisCommandArgs("EXPIRE"),
@@ -155,6 +157,17 @@ class Redis(RedisClient, RedisPrefixCheckMixin):
             shard_hint,
             key_generator=self.get_key,
         )
+
+    @cached_property
+    def elasticache(self) -> bool:
+        """
+        Check if this Redis instances is actually connected to AWS ElastiCache rather than Redis.
+
+        AWS ElastiCache is supposed to be API compatible with Redis, but there are some differences
+        that can cause issues. This property can be used to detect if we are connected to ElastiCache
+        and handle those differences.
+        """
+        return self.info().get("os") == "Amazon ElastiCache"
 
 
 class Pipeline(RedisPipeline, RedisPrefixCheckMixin):
